system_b/requirements.txt
Django==4.1.6
djangorestframework==3.14.0
drf-yasg==1.21.4
psycopg2==2.9.5
djangorestframework-simplejwt==5.2.2